demo/python/requirements.txt
numpy
matplotlib
roboticstoolbox-python
numpy-stl